USB: musb-new: omap2430: Add hooks for Peripheral support.
Borrowing from the updates to ti-musb.c, this glue layer checks
to see if the MUSB controller is gadget or host.  Going under
the assumpution there is only one MUSB controller, this only builds
either the host probe or the gadget probe based on the device tree
added by -u-boot.dtsi.  OTG would be preferred, but it doesn't
appear to be working yet.

Signed-off-by: Adam Ford <aford173@gmail.com>