usb: ehci-mx6: Create MISC USB device wrapper
In preparation to support DM_USB_GADGET, this creates a misc
device which binds the host.  This is necessary because
U_BOOT_DRIVER(usb_mx6) is currently hard coded to UCLASS_USB
and if we want to support UCLASS_USB_GADGET_GENERIC we'll need
to bind a separate device in the future.

Signed-off-by: Adam Ford <aford173@gmail.com>