ARM: dts: da850-evm: Pinmux MDIO pins for eth0
The Ethernet controller needs the MDIO pins enabled, but
the DM_MDIO driver and class are not quite ready yet, so as a
temporary fix, this places pinctrl for the mdio pins with
eth0.  Once the MDIO class and driver have been integrated,
this can be removed since the pin controller will use the
mdio node.

Fixes: 877ab242 ("ARM: davinci: da850: Manual pinmux only when
PINCTRL not available")

Signed-off-by: Adam Ford <aford173@gmail.com>