imx: Add support for i.MX8MM Beacon SOM 4GB variant
The Beacon Embedded has a new variant with a different LPDDR4 chip.
This new chip has 4GB of RAM.  Because the DDR timings are loaded
SPL, and there doesn't appear to be a way to determine which
chip timings are needed at runtime, a new config option is created to
enable the 4GB of memory and load the proper timings.

Signed-off-by: Adam Ford <aford173@gmail.com>