arm64: dts: imx8mn: Fix FSPI booting
When FSPI_CONF_HEADER is set, the binary needs to be built such
that there is a configuration file located at 0x400 and the start
of the file that would normally be flash.bin starts at 0x1000.
This used to be done properly until the device tree was converted to
nxp_imx8mimage.

Building these with the offsets built into the binman device tree
changes impacts how the actual image is built and the locations
of the various blobs aren't fetched properly and booting fails.

Fix this by building a standard image as if it were to boot from
eMMC or SD, then use that image as the input for a second image
which adds this FSPI header at 0x400, and then places the standard
image at offset 0x1000.  This also has the benefit of being able
to create both SD/eMMC image at the same time as a fspi image
called flexspi.bin.

Fixes: 37e50627efac ("ARM: dts: imx: Convert i.MX8M flash.bin image generation to binman")
Signed-off-by: Adam Ford <aford173@gmail.com>